+++ packages/at_onboarding_cli/lib/src/cli/auth_cli_args.dart
@@ -63,6 +63,7 @@ enum AuthCliCommand {
           ' program which has "rw" access to the "__manage" namespace.');
 
   const AuthCliCommand({this.usage = ''});
+
   final String usage;
 }
 
@@ -98,6 +99,7 @@ class AuthCliArgs {
   static const argNameAppNameRegex = 'arx';
   static const argNameDeviceNameRegex = 'drx';
   static const argNameMaxConnectAttempts = 'mca';
+  static const argNameExpiry = 'expiry';
 
   ArgParser get parser {
     return _aap;
@@ -282,6 +284,11 @@ class AuthCliArgs {
       help: 'The semi-permanent enrollment passcode to set for this atSign',
       mandatory: true,
     );
+    p.addOption(argNameExpiry,
+        abbr: 'e',
+        help:
+            'The duration for which the SPP remains active. The time duration can be passed as "2d,1h,10m,20s,999ms" for 2 days 1 hour 10 minutes 20 seconds 999 milliseconds',
+        mandatory: false);
     return p;
   }
 
@@ -295,6 +302,11 @@ class AuthCliArgs {
   @visibleForTesting
   ArgParser createOtpCommandParser() {
     ArgParser p = createSharedArgParser(hide: true);
+    p.addOption(argNameExpiry,
+        abbr: 'e',
+        help:
+            'The duration for which the OTP remains active. The time duration can be passed as "2d,1h,10m,20s,999ms" for 2 days 1 hour 10 minutes 20 seconds 999 milliseconds',
+        mandatory: false);
     return p;
   }
 
@@ -331,6 +343,11 @@ class AuthCliArgs {
           ' e.g. "buzz:rw,contacts:rw,__manage:rw"',
       mandatory: true,
     );
+    p.addOption(argNameExpiry,
+        abbr: 'e',
+        help:
+            'The duration for which the APKAM keys remains active. The time duration can be passed as "2d,1h,10m,20s,999ms" for 2 days 1 hour 10 minutes 20 seconds 999 milliseconds',
+        mandatory: false);
     return p;
   }