+++ packages/at_onboarding_cli/lib/src/onboard/at_onboarding_service_impl.dart
@@ -31,6 +31,7 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
   AtSignLogger logger = AtSignLogger('OnboardingCli');
   AtOnboardingPreference atOnboardingPreference;
   AtLookUp? _atLookUp;
+  final _maxActivationRetries = 5;
 
   /// The object which controls what types of AtClients, NotificationServices
   /// and SyncServices get created when we call [AtClientManager.setCurrentAtSign].
@@ -194,12 +195,9 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
   }
 
   @override
-  Future<at_auth.AtEnrollmentResponse> sendEnrollRequest(
-    String appName,
-    String deviceName,
-    String otp,
-    Map<String, String> namespaces,
-  ) async {
+  Future<at_auth.AtEnrollmentResponse> sendEnrollRequest(String appName,
+      String deviceName, String otp, Map<String, String> namespaces,
+      {Duration? apkamKeysExpiryDuration}) async {
     if (appName == null || deviceName == null) {
       throw AtEnrollmentException(
           'appName and deviceName are mandatory for enrollment');
@@ -207,11 +205,13 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
 
     at_auth.EnrollmentRequest newClientEnrollmentRequest =
         at_auth.EnrollmentRequest(
-      appName: appName,
-      deviceName: deviceName,
-      namespaces: namespaces,
-      otp: otp,
-    );
+            appName: appName,
+            deviceName: deviceName,
+            namespaces: namespaces,
+            otp: otp);
+    newClientEnrollmentRequest.apkamKeysExpiryDuration =
+        apkamKeysExpiryDuration;
+
     AtLookupImpl atLookUpImpl = AtLookupImpl(_atSign,
         atOnboardingPreference.rootDomain, atOnboardingPreference.rootPort);
     logger.finer('sendEnrollRequest: submitting enrollment request');
@@ -322,15 +322,44 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
     Duration retryInterval, {
     bool logProgress = true,
   }) async {
+    int retryAttempt = 1;
     while (true) {
       logger.info('Attempting pkam auth');
       if (logProgress) {
         stderr.write('Checking ... ');
       }
-      bool pkamAuthSucceeded = await _attemptPkamAuth(
-        atLookUp,
-        enrollmentIdFromServer,
-      );
+      bool pkamAuthSucceeded = false;
+      try {
+        // _attemptPkamAuth returns boolean value true when authentication is successful.
+        // Returns UnAuthenticatedException when authentication fails.
+        pkamAuthSucceeded = await atLookUp.pkamAuthenticate(
+            enrollmentId: enrollmentIdFromServer);
+      } on UnAuthenticatedException catch (e) {
+        // Error codes AT0401 and AT0026 indicate authentication failure due to unapproved enrollment. Retry until the enrollment is approved.
+        // The variable _pkamAuthSucceeded is false, allowing for PKAM authentication retries.
+        // Avoid checking "retryAttempt > _maxActivationRetries" here, as we want to continue retrying until enrollment is approved.
+        // The check for "retryAttempt > _maxActivationRetries" should only occur when the secondary server is unreachable due to network issues.
+        if (e.message.contains('error:AT0401') ||
+            e.message.contains('error:AT0026')) {
+          logger.info('Pkam auth failed: ${e.message}');
+        }
+        // Error code AT0025 represents Enrollment denied. Therefore, no need to retry; throw exception.
+        else if (e.message.contains('error:AT0025')) {
+          throw AtEnrollmentException(
+              'The enrollment: $enrollmentIdFromServer is denied');
+        }
+      } catch (e) {
+        String message =
+            'Exception occurred when authenticating the atSign: $_atSign caused by ${e.toString()}';
+        if (retryAttempt > _maxActivationRetries) {
+          message += ' Activation failed after $_maxActivationRetries attempts';
+          logger.severe(message);
+          rethrow;
+        }
+        logger
+            .severe('$message. Attempting to retry for $retryAttempt attempt');
+        retryAttempt++;
+      }
       if (pkamAuthSucceeded) {
         if (logProgress) {
           stderr.writeln(' approved.');
@@ -346,34 +375,6 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
         await Future.delayed(retryInterval); // Delay and retry
       }
     }
-  }
-
-  /// Try a single PKAM auth
-  Future<bool> _attemptPkamAuth(AtLookUp atLookUp, String enrollmentId) async {
-    try {
-      logger.finer('_attemptPkamAuth: Calling atLookUp.pkamAuthenticate');
-      var pkamResult =
-          await atLookUp.pkamAuthenticate(enrollmentId: enrollmentId);
-      logger.finer(
-          '_attemptPkamAuth: atLookUp.pkamAuthenticate returned $pkamResult');
-      if (pkamResult) {
-        return true;
-      }
-    } on UnAuthenticatedException catch (e) {
-      if (e.message.contains('error:AT0401') ||
-          e.message.contains('error:AT0026')) {
-        logger.info('Pkam auth failed: ${e.message}');
-        return false;
-      } else if (e.message.contains('error:AT0025')) {
-        throw AtEnrollmentException('enrollment denied');
-      }
-    } catch (e) {
-      logger.shout('Unexpected exception: $e');
-      rethrow;
-    } finally {
-      logger.finer('_attemptPkamAuth: complete');
-    }
-    return false;
   }
 
   ///write newly created encryption keypairs into atKeys file